+++ winpr/libwinpr/sysinfo/test/TestGetNativeSystemInfo.c
@@ -14,7 +14,7 @@ int TestGetNativeSystemInfo(int argc, char* argv[])
 	printf("\tdwPageSize: 0x%08lX\n", sysinfo.dwPageSize);
 	printf("\tlpMinimumApplicationAddress: %p\n", sysinfo.lpMinimumApplicationAddress);
 	printf("\tlpMaximumApplicationAddress: %p\n", sysinfo.lpMaximumApplicationAddress);
-	printf("\tdwActiveProcessorMask: 0x%08llX\n", sysinfo.dwActiveProcessorMask);
+	printf("\tdwActiveProcessorMask: 0x%08llX\n", (unsigned long long)sysinfo.dwActiveProcessorMask);
 	printf("\tdwNumberOfProcessors: %ld\n", sysinfo.dwNumberOfProcessors);
 	printf("\tdwProcessorType: %ld\n", sysinfo.dwProcessorType);
 	printf("\tdwAllocationGranularity: %ld\n", sysinfo.dwAllocationGranularity);
